Move home/plus setting to address bar setting section.
