+++ firefox-ios/Client/Frontend/Settings/AppearanceSettings/AppearanceSettingsView.swift
@@ -4,7 +4,6 @@
 
 import SwiftUI
 import Common
-import Shared
 
 /// Child settings pages appearance actions
 protocol AppearanceSettingsDelegate: AnyObject {
@@ -20,16 +19,10 @@ struct AppearanceSettingsView: View, FeatureFlaggable {
     @Environment(\.themeManager)
     var themeManager
 
-    var prefs: Prefs
-
     @State private var currentTheme: Theme?
 
     var shouldShowPageZoom: Bool {
         return featureFlags.isFeatureEnabled(.defaultZoomFeature, checking: .buildOnly)
-    }
-
-    var shouldShowNavigationBarConfig: Bool {
-        return featureFlags.isFeatureEnabled(.toolbarMiddleButtonCustomization, checking: .buildOnly)
     }
 
     /// Compute the theme option to display in the ThemeSelectionView.
@@ -41,15 +34,6 @@ struct AppearanceSettingsView: View, FeatureFlaggable {
         } else {
             return themeManager.getUserManualTheme() == .light ? .light : .dark
         }
-    }
-
-    var selectedMiddleButtonType: NavigationBarMiddleButtonType {
-        if let rawValue = prefs.stringForKey(PrefsKeys.Settings.navigationToolbarMiddleButton),
-           let selectedButton = NavigationBarMiddleButtonType(rawValue: rawValue) {
-            return selectedButton
-        }
-
-        return .newTab
     }
 
     private var viewBackground: Color {
@@ -85,12 +69,6 @@ struct AppearanceSettingsView: View, FeatureFlaggable {
                     }
                 }
 
-                if shouldShowNavigationBarConfig {
-                    NavigationToolbarSection(theme: currentTheme,
-                                             selectedOption: selectedMiddleButtonType,
-                                             onChange: updateMiddleNavigationToolbarButton,
-                                             cornerRadius: UX.cornerRadius)
-                }
                 Spacer()
             }
         }
@@ -192,40 +170,5 @@ struct AppearanceSettingsView: View, FeatureFlaggable {
         } else {
             // TODO(FXIOS-11584): Add telemetry here
         }
-    }
-
-    // MARK: NavigationToolbarSection
-
-    private struct NavigationToolbarSection: View {
-        let theme: Theme?
-        let selectedOption: NavigationBarMiddleButtonType
-        let onChange: (NavigationBarMiddleButtonType) -> Void
-        let cornerRadius: CGFloat
-
-        var body: some View {
-            GenericSectionView(
-                theme: theme,
-                title: .Settings.Appearance.NavigationToolbar.SectionHeader,
-                description: .Settings.Appearance.NavigationToolbar.SectionDescription,
-                identifier: AccessibilityIdentifiers.Settings.Appearance.navigationToolbarSectionTitle
-            ) {
-                NavigationBarMiddleButtonSelectionView(
-                    theme: theme,
-                    selectedMiddleButton: selectedOption,
-                    onSelected: onChange)
-                .modifier(SectionStyle(theme: theme, cornerRadius: cornerRadius))
-            }
-        }
-    }
-
-    /// Updates the middle button in navigation toolbar based on the user's selection.
-    /// - Parameter selectedOption: The selected theme option from ThemeSelectionView.
-    private func updateMiddleNavigationToolbarButton(to selectedOption: NavigationBarMiddleButtonType) {
-        prefs.setString(selectedOption.rawValue, forKey: PrefsKeys.Settings.navigationToolbarMiddleButton)
-
-        let action = ToolbarAction(middleButton: selectedOption,
-                                   windowUUID: windowUUID,
-                                   actionType: ToolbarActionType.navigationMiddleButtonDidChange)
-        store.dispatchLegacy(action)
     }
 }
